requirements_live.txt
# OKX Ultimate 75% Trading Bot - Live Requirements
# Core trading libraries
websocket-client>=1.6.0
pandas>=1.5.0
numpy>=1.21.0
aiohttp>=3.8.0
asyncio
requests>=2.31.0

# Data analysis and visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dashboard and UI
streamlit>=1.28.0

# Database for logging
sqlite3

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: Enhanced analysis
scikit-learn>=1.3.0
scipy>=1.11.0

# Performance optimization
numba>=0.57.0

# Live trading with OKX
okx>=1.0.0
ccxt>=4.1.0

# Async operations
asyncio-mqtt>=0.14.0
websockets>=11.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.0.0

# Cython for performance
cython>=3.0.0

# Trade notifications
plyer>=2.1.0
